+++ Lab 5/src/main/java/com/example/lab5/service/TeacherService.java
@@ -67,6 +67,7 @@ public class TeacherService {
     }
 
     public void delete(UUID id) {
+        if (!teacherRepository.existsById(id)) throw new ItemNotFoundException(ItemNotFoundException.ItemType.TEACHER, id);
         teacherRepository.deleteById(id);
     }
